+++ src/main/java/carracing/model/car/strategy/MovingStrategy.java
@@ -1,4 +1,4 @@
-package carracing.model;
+package carracing.model.car.strategy;
 
 @FunctionalInterface
 public interface MovingStrategy {
