+++ src/main/java/carracing/model/car/strategy/RandomStrategy.java
@@ -1,4 +1,4 @@
-package carracing.model;
+package carracing.model.car.strategy;
 
 import java.util.Random;
